+++ compiler/src/test/resources/FieldsInjection.java
@@ -1,0 +1,13 @@
+import com.contentful.vault.BaseFields;
+
+public final class Test$Model$Fields extends BaseFields {
+  public static final String FOO = "foo";
+
+  public static final String BAR = "bar";
+
+  public static final String BAZ = "baz";
+
+  public static final String BAZ2 = "baz2";
+
+  public static final String B_A_Z_123 = "b_a_z_123";
+}