+++ core/src/main/java/com/contentful/vault/BaseFields.java
@@ -1,0 +1,9 @@
+package com.contentful.vault;
+
+public abstract class BaseFields {
+  public static final String REMOTE_ID = "remote_id";
+
+  public static final String CREATED_AT = "created_at";
+
+  public static final String UPDATED_AT = "updated_at";
+}